Complete uta final ranking generation

+++ backend/src/main/kotlin/pl/poznan/put/decision_support/decisionmethods/model/domain/UTA/response/UTAResponse.kt
@@ -2,8 +2,8 @@ package pl.poznan.put.decision_support.decisionmethods.model.domain.UTA.response
 
 data class UTAResponse(
     val Kendall: Kendall? = null,
-    val errors: List<Int>,
-    val optimum: List<Int>,
+    val errors: List<Float>,
+    val optimum: List<Float>,
     val overallValues: List<Float>,
     val ranks: List<Int>,
     val valueFunctions: Map<String, List<List<Double>>>? = null
